AE-2156: Move primary toimenpide highlighting related components to their own directory
